bin/requirements.txt
pyqt==5.15.6
pytube==12.1.0